% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.triangle.R
\name{as.triangle}
\alias{as.triangle}
\title{as.triangle}
\usage{
as.triangle(triangleDT, valueCol = "Transactions",
  descriptiveHeaders = TRUE, idCols = "Cohort")
}
\arguments{
\item{triangleDT}{A triangle in tall, data.table format}

\item{valueCol}{Name of the column to convert to triangular format}

\item{descriptiveHeaders}{Should headers be descriptive?}
}
\description{
Convert a single triangle from tall format to triangular format
}
\details{
Returns a matrix object with rows representing origin periods and columns representing valuation ages
}
\examples{
library(data.table)

set.seed(2357)
transactions <- sample_transactions(10)
triangles <- make_triangles(transactions, format="tall")
as.triangle(triangles, valueCol="ActiveCustomers")
as.triangle(triangles, valueCol="Transactions")
}
