% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulate_transactions.R
\name{cumulate_transactions}
\alias{cumulate_transactions}
\title{Cumulate Transactions}
\usage{
cumulate_transactions(transactions, colCustomerID = "CustomerID",
  colTransactionDate = "TransactionDate", financialCols = NULL)
}
\arguments{
\item{transactions}{A data.table of transactions with a column for CustomerID and TransactionDate}

\item{colCustomerID}{Name of the CustomerID column}

\item{colTransactionDate}{Name of the TransactionDate column}

\item{financialCols}{Character vector corresponding to the name of financial columns in \code{transactions}}
}
\description{
Aggregate transactions monthly and measure cumulative financial values as well as cumulative transaction count per customer
}
\details{
Returns a data.table object with columns {[CustomerID], FirstValuationDate, ValuationDate, Transactions.cmltv} and
{Financial1.cmltv, Financial2.cmltv, ...} for all specified financial columns in the transactions dataset
}
\examples{
library(data.table)

set.seed(2016)
transactions <- sample_transactions(10)
cumulate_transactions(transactions)
cumulate_transactions(transactions, financialCols=c("Amount"))
}

