% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customer_cohorts.R
\name{customer_cohorts}
\alias{customer_cohorts}
\title{Customer Cohorts}
\usage{
customer_cohorts(transactions.cmltv, customers, colCohort,
  colCustomerID = "CustomerID", colTransactionDate = "TransactionDate",
  colsFinancial = "auto", allCohort = TRUE, format = "triangular",
  minLeftOrigin = min(transactions.cmltv$FirstValuationDate),
  originLength = 12, rowDev = 12, colDev = 12,
  lastValuationDate = max(transactions.cmltv$ValuationDate),
  fromMinLeftOrigin = TRUE, initialAge = originLength, verbose = FALSE)
}
\arguments{
\item{transactions.cmltv}{A data.table of cumulative transaction valuations (result of calling cumulate_transactions())}

\item{customers}{A data.table of customers (probably the result of calling get_customers())}

\item{colCohort}{Name of the column for which to group cohorts}

\item{colCustomerID}{Name of the CustomerID column}

\item{colTransactionDate}{Name of the TransactionDate column}

\item{colsFinancial}{Character vector corresponding to the name of financial columns in \code{transactions.cmltv}}

\item{allCohort}{Triangles be returned representing the combined sum of all cohorts?}

\item{format}{How should the triangles be returned? Either "tall" (a data.table) or "triangular" (a list of data.tables)}

\item{minLeftOrigin}{See ?triangle_skeleton}

\item{originLength}{See ?triangle_skeleton}

\item{rowDev}{See ?triangle_skeleton}

\item{colDev}{See ?triangle_skeleton}

\item{lastValuationDate}{See ?triangle_skeleton}

\item{fromMinLeftOrigin}{See ?triangle_skeleton}

\item{initialAge}{See ?triangle_skeleton}

\item{verbose}{Should progress details be displayed?}
}
\description{
Generate customer level triangles, i.e. triangles per customer-level feature (e.g. first product purchased)
}
\details{
Returns a list of triangles (possibly a list of lists, if format="triangular"), where the top level list elements correspond
to the unique values of the colCohort in the customers dataset
}
\examples{
library(data.table)

set.seed(2357)
transactions <- sample_transactions(3, minDate=as.Date("2010-1-1"), maxDate=as.Date("2015-12-31"))
customers <- get_customers(transactions, colsFinancial="Amount", catCols="Product")
transactions.cmltv <- cumulate_transactions(transactions, colsFinancial="Amount")
customer_cohorts(transactions.cmltv, customers, colCohort="Product.First")
}

