% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eom.R
\name{eom}
\alias{eom}
\title{End of Month}
\usage{
eom(dates)
}
\arguments{
\item{dates}{A vector of dates}
}
\description{
Return the last day of the month for each date in a vector of dates
}
\details{
Returns the last day of the month for each date in a vector of dates
}
\examples{
eom(as.Date("2016-2-14"))
eom(seq.Date(as.Date("2015-1-1"), as.Date("2015-12-31"), by="month"))
}
