% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fom.R
\name{fom}
\alias{fom}
\title{End of Month}
\usage{
fom(dates, p = as.POSIXlt(dates))
}
\arguments{
\item{dates}{A vector of dates}
}
\description{
Return the first day of the month for each date in a vector of dates
}
\details{
Returns the first day of the month for each date in a vector of dates
}
\examples{
fom(as.Date("2016-2-14"))
fom(seq.Date(as.Date("2015-1-1"), as.Date("2015-12-31"), by="month"))
}
