% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_customers.R
\name{get_customers}
\alias{get_customers}
\title{Get Customers}
\usage{
get_customers(transactions, colCustomerID = "CustomerID",
  colTransactionDate = "TransactionDate", financialCols = NULL,
  catCols = NULL)
}
\arguments{
\item{nCusts}{How many customers?}

\item{sdTransactions}{The number of transactions per customer = \code{abs(ceiling(rnorm(n=1, sd=sdTransactions)))}}

\item{sdAmount}{The amount of each transaction = \code{abs(round(rnorm(n=1, sd=sdAmount),2))}}

\item{minDate}{The minimum possible date a transaction could occur}

\item{maxDate}{The maximum possible date a transaction could occur}

\item{products}{A character vector of potential product names}
}
\description{
Build a customers dataset from a transactions dataset with required columns {CustomerID, TransactionDate} (or alternate names)
}
\details{
Returns a data.table object with one row per unique CustomerID
}
\examples{
library(data.table)

set.seed(2016)
transactions <- sample_transactions(10)
get_customers(transactions)
get_customers(transactions, financialCols=c("Amount"), catCols=c("Product"))
}

