% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_transactions.R
\name{sample_transactions}
\alias{sample_transactions}
\title{Sample Transactions}
\usage{
sample_transactions(nCusts, sdTransactions = 10, sdAmount = 10,
  minDate = as.Date("2014-1-1"), maxDate = as.Date("2015-12-31"),
  products = LETTERS)
}
\arguments{
\item{nCusts}{How many customers?}

\item{sdTransactions}{The number of transactions per customer = \code{abs(ceiling(rnorm(n=1, sd=sdTransactions)))}}

\item{sdAmount}{The amount of each transaction = \code{abs(round(rnorm(n=1, sd=sdAmount),2))}}

\item{minDate}{The minimum possible date a transaction could occur}

\item{maxDate}{The maximum possible date a transaction could occur}

\item{products}{A character vector of potential product names}
}
\description{
Generate a sample transactions dataset such that the simulated data has a (somewhat) realistic pattern of customer churn
}
\details{
Returns a data.table object with columns {TransactionID, TransactionDate, CustomerID, Amount, Product}
}
\examples{
library(data.table)

set.seed(2016)
sample_transactions(nCusts=10)
sample_transactions(nCusts=10, sdTransactions=5, sdAmount=1000)
sample_transactions(nCusts=10, minDate=as.Date("2000-1-1"), maxDate=as.Date("2016-12-31"))
sample_transactions(nCusts=10, products=c("baseball", "football", "soccerball"))
}

