% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transaction_cohorts.R
\name{transaction_cohorts}
\alias{transaction_cohorts}
\title{Transaction Cohorts}
\usage{
transaction_cohorts(transactions, colCohort, colCustomerID = "CustomerID",
  colTransactionDate = "TransactionDate", colsFinancial = NULL,
  firstTransactionWRTCohort = FALSE, allCohort = TRUE,
  format = "triangular",
  minLeftOrigin = as.Date(paste0(min(year(transactions[[colTransactionDate]])),
  "-1-1")), originLength = 12, rowDev = 12, colDev = 12,
  lastValuationDate = max(transactions[[colTransactionDate]]),
  fromMinLeftOrigin = TRUE, initialAge = originLength, verbose = FALSE)
}
\arguments{
\item{transactions}{A data.table of transactions with a column for CustomerID, TransactionDate, and Cohort}

\item{colCohort}{Name of the column for which to group cohorts}

\item{firstTransactionWRTCohort}{If TRUE, a single customer can be mapped to many origins - each based on his first transaction
within each cohort. If FALSE, each customer has a fixed origin based on his first overall transaction}

\item{allCohort}{Triangles be returned representing the combined sum of all cohorts?}

\item{format}{"tall" or "triangular"}

\item{minLeftOrigin}{See ?triangle_skeleton}

\item{originLength}{See ?triangle_skeleton}

\item{rowDev}{See ?triangle_skeleton}

\item{colDev}{See ?triangle_skeleton}

\item{lastValuationDate}{See ?triangle_skeleton}

\item{fromMinLeftOrigin}{See ?triangle_skeleton}

\item{initialAge}{See ?triangle_skeleton}

\item{verbose}{Should progress details be displayed?}
}
\description{
Generate triangles from transactions relative to a specified cohort column
}
\details{
Returns a list of triangles (possibly a list of lists, if \code{format="triangular"}), where the top level list elements
correspond to the unique values of the \code{colCohort} in the transactions dataset
}
\examples{
library(data.table)

set.seed(2357)
transactions <- sample_transactions(10)
transaction_cohorts(transactions, colCohort="Product")
}

