% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangle_skeleton.R
\name{triangle_skeleton}
\alias{triangle_skeleton}
\title{Triangle Skeleton}
\usage{
triangle_skeleton(minLeftOrigin, fromMinLeftOrigin = FALSE,
  originLength = 12, rowDev = 12, colDev = 12,
  initialAge = originLength, lastValuationDate = Sys.Date())
}
\arguments{
\item{minLeftOrigin}{When should the first origin period (potentially) start? (Depends on \code{fromMinLeftOrigin})}

\item{fromMinLeftOrigin}{Should the first origin period start on exactly \code{minLeftOrigin}? If \code{FALSE} the last
valuation date will be exactly \code{lastValuationDate}}

\item{originLength}{The length of each origin period in months}

\item{rowDev}{Number of months between rows}

\item{colDev}{Number of months between columns}

\item{initialAge}{At what age should the first valuation period start?}

\item{lastValuationDate}{When whould the last valuation date (potentially) be? (Depends on \code{fromMinLeftOrigin})}
}
\description{
Generate the dates needed to build a triangle analysis
}
\details{
Returns a data.table with columns {LeftOrigin, RightOrigin, ValuationDate, Age}
}
\examples{
triangle_skeleton(as.Date("2014-1-1"))
eom(seq.Date(as.Date("2015-1-1"), as.Date("2015-12-31"), by="month"))
}

