% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_triangles.R
\name{write_triangles}
\alias{write_triangles}
\title{Write Triangles}
\usage{
write_triangles(triangles, filename = "triangles.csv")
}
\arguments{
\item{triangles}{Triangles in tall or a list of them in triangular format, or a list of either of those}

\item{filename}{filename (optionally with path) determining where to save triangles. DEFAULT is triangles.csv inside the
current working directory}
}
\description{
Method to write many formats of triangles to a CSV file
}
\details{
Writes triangles to a CSV file
}
\examples{
library(data.table)

set.seed(2016)
transactions <- sample_transactions(10)
tris <- make_triangles(transactions)
write_triangles(tris)
}
