% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customer_cohorts.R
\name{customer_cohorts}
\alias{customer_cohorts}
\title{Make Triangles}
\usage{
customer_cohorts(customers, transactions.cmltv, colCohort,
  colCustomerID = "CustomerID", colsFinancial = "auto",
  colTransactionDate = "TransactionDate", allCohort = TRUE,
  format = "triangular",
  minLeftOrigin = min(transactions.cmltv$FirstValuationDate),
  originLength = 12, rowDev = 12, colDev = 12,
  lastValuationDate = max(transactions.cmltv$ValuationDate),
  fromMinLeftOrigin = TRUE, initialAge = originLength, verbose = FALSE)
}
\arguments{
\item{transactions.cmltv}{A data.table of cumulative transaction valuations (result of calling cumulate_transactions())}

\item{colsFinancial}{What financial columns in \code{transactions.cmltv} should generate triangles? Default="auto" guesses}

\item{format}{How should the triangles be returned? Either "tall" (a data.table) or "triangular" (a list of data.tables)}

\item{minLeftOrigin}{See ?triangle_skeleton}

\item{originLength}{See ?triangle_skeleton}

\item{rowDev}{See ?triangle_skeleton}

\item{colDev}{See ?triangle_skeleton}

\item{lastValuationDate}{See ?triangle_skeleton}

\item{fromMinLeftOrigin}{See ?triangle_skeleton}

\item{initialAge}{See ?triangle_skeleton}

\item{verbose}{Should progress details be displayed?}
}
\description{
Convert triangles in tall format to a list of triangles in triangular format
}
\details{
Returns a list of data.table objects
}
\examples{
library(data.table)

set.seed(2357)
transactions <- sample_transactions(3, minDate=as.Date("2010-1-1"), maxDate=as.Date("2015-12-31"))
customers <- get_customers(transactions, colsFinancial="Amount", catCols="Product")
transactions.cmltv <- cumulate_transactions(transactions, colsFinancial="Amount")
customer_cohorts(transactions.cmltv, customers)
}

