% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_customers.R
\name{get_customers}
\alias{get_customers}
\title{Get Customers}
\usage{
get_customers(transactions, colCustomerID = "CustomerID",
  colTransactionDate = "TransactionDate", colsFinancial = NULL,
  colsCategory = NULL)
}
\arguments{
\item{transactions}{A data.table of transactions with a column for CustomerID and TransactionDate}

\item{colCustomerID}{Name of the CustomerID column}

\item{colTransactionDate}{Name of the TransactionDate column}

\item{colsFinancial}{Character vector corresponding to the name of financial columns in \code{transactions}}

\item{colsCategory}{Character vector corresponding to the name of categorical columns in \code{transactions}}
}
\description{
Build a customers dataset from a transactions dataset with required columns {CustomerID, TransactionDate} (or alternate names)
}
\details{
Returns a data.table object with one row per unique CustomerID
}
\examples{
library(data.table)

set.seed(2016)
transactions <- sample_transactions(10)
get_customers(transactions)
get_customers(transactions, colsFinancial=c("Amount"), colsCategory=c("Product"))
}
