% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_triangles.R
\name{make_triangles}
\alias{make_triangles}
\title{Make Triangles}
\usage{
make_triangles(transactions.cmltv, format = "triangular",
  minLeftOrigin = as.Date(paste0(min(year(transactions.cmltv$FirstValuationDate)),
  "-1-1")), originLength = 12, rowDev = 12, colDev = 12,
  lastValuationDate = max(transactions.cmltv$ValuationDate),
  fromMinLeftOrigin = TRUE, initialAge = originLength,
  financialCols = "auto", verbose = FALSE)
}
\arguments{
\item{transactions.cmltv}{A data.table of cumulative transaction valuations (result of calling cumulate_transactions())}

\item{format}{How should the triangles be returned? Either "tall" (a data.table) or "triangular" (a list of data.tables)}

\item{minLeftOrigin}{See ?triangle_skeleton}

\item{originLength}{See ?triangle_skeleton}

\item{rowDev}{See ?triangle_skeleton}

\item{colDev}{See ?triangle_skeleton}

\item{lastValuationDate}{See ?triangle_skeleton}

\item{fromMinLeftOrigin}{See ?triangle_skeleton}

\item{initialAge}{See ?triangle_skeleton}

\item{financialCols}{What financial columns in \code{transactions.cmltv} should generate triangles? Default="auto" guesses}
}
\description{
Convert triangles in tall format to a list of triangles in triangular format
}
\details{
Returns a list of data.table objects
}
\examples{
library(data.table)

set.seed(2357)
transactions <- sample_transactions(3, minDate=as.Date("2010-1-1"), maxDate=as.Date("2015-12-31"))
transactions.cmltv <- cumulate_transactions(transactions, financialCols="Amount")
make_triangles(transactions.cmltv)  # guess the financial columns
make_triangles(transactions.cmltv, financialCols=c("Transactions.cmltv"))  # specify the financial columns
}

