% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trall_to_triangular.R
\name{tall_to_triangular}
\alias{tall_to_triangular}
\title{Tall to Triangular}
\usage{
tall_to_triangular(triangleDT, descriptiveHeaders = TRUE)
}
\arguments{
\item{triangleDT}{Triangles in tall, data.table format}

\item{triangleDT}{Should headers be descriptive?}
}
\description{
Convert triangles in tall format to a list of triangles in triangular format
}
\details{
Returns a list of data.table objects
}
\examples{
library(data.table)

set.seed(2357)
transactions <- sample_transactions(10)
triangles <- make_triangles(transactions, format="tall")
tall_to_triangular(triangles)
}
